As discussed w/ Zibin. Updated TransactionModel to better fit payments. Updated Tests as well. NOTE: Requests won't work atm until the User model gets merged into main and I can create some dumby users

+++ src/main/java/com/revlearn/team1/model/PaymentModel.java
@@ -23,7 +23,7 @@ public class PaymentModel
     private String firstName;
     private String lastName;
     private String email;
-    private int cardNumber;
+    private long cardNumber;
     private String expirationDate;
     private int cvc;
     private int zip;
